Update numpy version constraints in requirements.txt (oversight causing a dependency conflict that was unresolvable)
This more specifically is concerning OpenCV-Python: the version specified causes a conflict, which it required NumPy versions <2.3.0 and >=2.0.0 (also assuming that Python version >=3.9) while the requested NumPy version was 2.3.2, which caused a contradiction to become unresolvable. NumPy version constraints were changed to allow more flexibility.

Tested it, and it works.